requirements.txt
gensim
nltk
scikit-learn
pandas
numpy
matplotlib
pymongo[srv]
python-dotenv
pyjwt
flask